Adds option to use today as the start date for calculations.
When used with another date argument, start will be today and the other
argument will be used as the end date.

We also had to set a default output date format because we can no longer
rely on there being a start date to match the format of. YYYY-MM-DD was
selected because it seemed more readable.